Added get animal type by id

+++ AnimalChipization/AnimalChipization.Services/Services/Interfaces/ILocationService.cs
@@ -5,7 +5,7 @@ namespace AnimalChipization.Services.Services.Interfaces;
 
 public interface ILocationService
 {
-    Task<Location?> GetByIdAsync(int locationId);
+    Task<Location?> GetByIdAsync(long locationId);
     Task CreateAsync(Location model);
     Task<Location> UpdateAsync(UpdateLocationModel model);
 }
